Force float32 data type in correlation coefficient computation.

+++ dyfunconn/fc/crosscorr.py
@@ -20,7 +20,7 @@ def crosscorr(data, fb, fs, pairs=None):
     n_channels, n_samples = np.shape(data)
     filtered, _, _ = analytic_signal(data, fb, fs)
 
-    r = np.zeros([n_channels, n_channels])
+    r = np.zeros([n_channels, n_channels], dtype=np.float32)
 
     for i in range(n_channels):
         for ii in range(n_channels):
