+++ dyfunconn/fc/partcorr.py
@@ -17,7 +17,7 @@ def partcorr(data, fb, fs, pairs=None):
     n_channels, n_samples = np.shape(data)
     filtered, _, _ = analytic_signal(data, fb, fs)
 
-    P_corr = np.zeros((n_channels, n_channels), dtype=np.float)
+    P_corr = np.zeros((n_channels, n_channels), dtype=np.float32)
 
     r = np.corrcoef(filtered)
     rinv = np.linalg.inv(r)
